Almost completed seeding, user roles are not done yet for seeding portion, index pages are done

+++ ItemInsight/Server/Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -370,8 +370,8 @@ namespace ItemInsight.Server.Data.Migrations
                     b.Property<int>("ProductId")
                         .HasColumnType("int");
 
-                    b.Property<int?>("Rating")
-                        .HasColumnType("int");
+                    b.Property<double?>("Rating")
+                        .HasColumnType("float");
 
                     b.Property<string>("UpdatedBy")
                         .HasColumnType("nvarchar(max)");
